43 put dmi point in db

+++ server/src/main/java/com/project/aau/sw3/p3/model/TotalPrecipitation.java
@@ -7,7 +7,7 @@ public class TotalPrecipitation {
     private String dataType;
     private List<String> axisNames;
     private List<Integer> shape;
-    private List<Float> values;
+    private List<Double> values;
 
     //getters and setters
 
@@ -43,11 +43,11 @@ public class TotalPrecipitation {
         this.shape = shape;
     }
 
-    public List<Float> getValues() {
+    public List<Double> getValues() {
         return values;
     }
 
-    public void setValues(List<Float> values) {
+    public void setValues(List<Double> values) {
         this.values = values;
     }
 }
